prettyfying

+++ start.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
+restart=$1
 
-rm -rf disk
-python setup.py
+if [ "$restart" == "-r" ]; then
+	rm -rf disk
+	python setup.py
+fi
+
 python user_api.py
